Move DefaultModuleBehavior into it's own package, and fix warnings.
