+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modules/RobotModule.java
@@ -1,5 +1,7 @@
 package org.firstinspires.ftc.teamcode.modules;
 
+import org.firstinspires.ftc.teamcode.modules.defaultmodulebehavior.DefaultModuleBehavior;
+
 public interface RobotModule {
     @DefaultModuleBehavior
     default void defaultModuleBehavior() {}
